Backend/UserService/UserCreate/requirements.txt
bcrypt === 4.0.1
pymongo== 4.3.3
pyjwt == 2.6.0

